requirements.txt
numpy
sentence-transformers
PyMuPDF
nltk
faiss-cpu
tf-keras
fastapi
uvicorn
boto3
python-multipart
python-dotenv
langchain
sentence-transformers 
transformers
accelerate
huggingface_hub
langchain-community
